// apps/server/src/modules/common-cartridge/common-cartridge-client/lesson-client/lessons-api-client/models/component-etherpad-props-impl.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ComponentEtherpadPropsImpl
 */
export interface ComponentEtherpadPropsImpl {
    /**
     * description of a Etherpad component
     * @type {string}
     * @memberof ComponentEtherpadPropsImpl
     */
    'description': string;
    /**
     * title of a Etherpad component
     * @type {string}
     * @memberof ComponentEtherpadPropsImpl
     */
    'title': string;
    /**
     * url of a Etherpad component
     * @type {string}
     * @memberof ComponentEtherpadPropsImpl
     */
    'url': string;
}